Add user "inheritance"
Execute the ./kankyo script to use container, this will evaluate the
current user and push the same UID and GID into the container. This
allows sane editing of files without worry of modifying perms. Also add
the docker user to the user so that `docker` commands can be used.

Also slight cleanup of vimrc.